requirements.txt
confluent-kafka==2.3.0
python-dotenv==1.0.0
faker==20.1.0
typing-extensions==4.8.0
